Changed the way the monte-carlo integration is done.
Due to instability in the monte carlo integration, I changed from
control variate to importance sampling from t-distribution, using the
mode and hessian as the mean and covariance matrix. I also changed the
way the baseline is calculated to direct integration. There is still a
problem in the exclude case where it breaks for some parameter
combination, probably due to it not being continuous.